requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
gradio>=4.0.0
python-dotenv